Пробуем сделать тесты на Git

+++ src/main/java/com/example/servicediplom/dto/event/NewEventDto.java
@@ -12,7 +12,7 @@ public class NewEventDto {
     @Size(min = 20, max = 2000)
     String annotation;
 
-    int category;
+    Long category;
     @Size(min = 20,max = 7000)
     String description;
